jjp-155: rename dto packages
