+++ auth_app/src/main/java/com/blankfactor/auth/entity/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.blankfactor.auth.entity.dto.incoming;
+package com.blankfactor.auth.entity.dto.request;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Builder;
