+++ auth_app/src/main/java/com/blankfactor/auth/entity/dto/request/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.blankfactor.auth.entity.dto.incoming;
+package com.blankfactor.auth.entity.dto.request;
 
 import jakarta.validation.constraints.*;
 import lombok.*;
