+++ auth_app/src/test/java/com/blankfactor/auth/controller/AuthControllerTests.java
@@ -2,11 +2,11 @@ package com.blankfactor.auth.controller;
 
 import com.blankfactor.auth.entity.Role;
 import com.blankfactor.auth.entity.User;
-import com.blankfactor.auth.entity.dto.outgoing.RegisterResponse;
-import com.blankfactor.auth.entity.dto.outgoing.VerifyResponse;
-import com.blankfactor.auth.entity.dto.incoming.LoginRequest;
-import com.blankfactor.auth.entity.dto.incoming.RegisterRequest;
-import com.blankfactor.auth.entity.dto.incoming.VerifyRequest;
+import com.blankfactor.auth.entity.dto.response.RegisterResponse;
+import com.blankfactor.auth.entity.dto.response.VerifyResponse;
+import com.blankfactor.auth.entity.dto.request.LoginRequest;
+import com.blankfactor.auth.entity.dto.request.RegisterRequest;
+import com.blankfactor.auth.entity.dto.request.VerifyRequest;
 import com.blankfactor.auth.exception.custom.InvalidCredentialsException;
 import com.blankfactor.auth.exception.custom.InvalidInformRequestException;
 import com.blankfactor.auth.exception.custom.PasswordsDoNotMatchException;
